bb_back/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
pydantic==1.8.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
psycopg2==2.9.5
django-environ==0.9.0
environ==1.0
django-rest-swagger==2.2.0
drf_yasg